backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
python-dotenv
requests
numpy
streamlit
bcrypt
pandas